// src/model/StockProfite.ts
export interface StockProfite {
    ST_PROFIT_ID?: number;
    TRAND_ID?: number;
    TRANS_DATE?: Date;
    ZAKAH_DATE?: Date;
    STOCK_SYMBOL?: string;
    CCOMPANY_NAME?: string;
    NO_STOCKS?: number;
    SPOT_PRICE_OZ?: number;
    MARKET_PRICE?: number;
    SALES_TAX?: number;
    TOT_MARKET_SALE_TAX?: number;
    TRANS_COST?: number;
    VALUE_BASED_CURR_PRICE?: number;
    PROFIT_VALUE?: number;
    PROFIT_PERCENT?: number;
    ZAKAH_VALUE?: number;
}